Split code in files finalized
